Move conftest fixtures
